Move files to services project
