+++ src/Services/Events/Processors/EventProcessorProvider.cs
@@ -1,4 +1,3 @@
-using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading;
