+++ src/Services/Services.csproj
@@ -6,13 +6,12 @@
     </PropertyGroup>
 
     <ItemGroup>
-        <PackageReference Include="Microsoft.AspNetCore.App" />
+        <FrameworkReference Include="Microsoft.AspNetCore.App" />
         <PackageReference Include="Newtonsoft.Json" Version="12.0.3" />
         <PackageReference Include="System.Threading.Channels" Version="4.7.1" />
     </ItemGroup>
 
     <ItemGroup>
-        <ProjectReference Include="..\Clients\Clients.csproj" />
         <ProjectReference Include="..\Models\Models.csproj" />
     </ItemGroup>
