+++ src/SlackApp.sln
@@ -7,8 +7,6 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Api", "Api\Api.csproj", "{6
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Models", "Models\Models.csproj", "{FA1FAC5E-3617-4FDD-85E3-C184BAA06593}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Clients", "Clients\Clients.csproj", "{AFFE7792-EB58-493B-9098-791C8AF656FC}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Tests", "Tests\Tests.csproj", "{47D18D7C-1498-4318-A466-5A5CE6B80327}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Services", "Services\Services.csproj", "{1C4E6C13-57A5-4882-AFEA-058490A17A87}"
@@ -50,18 +48,6 @@ Global
 		{FA1FAC5E-3617-4FDD-85E3-C184BAA06593}.Release|x64.Build.0 = Release|Any CPU
 		{FA1FAC5E-3617-4FDD-85E3-C184BAA06593}.Release|x86.ActiveCfg = Release|Any CPU
 		{FA1FAC5E-3617-4FDD-85E3-C184BAA06593}.Release|x86.Build.0 = Release|Any CPU
-		{AFFE7792-EB58-493B-9098-791C8AF656FC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{AFFE7792-EB58-493B-9098-791C8AF656FC}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{AFFE7792-EB58-493B-9098-791C8AF656FC}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{AFFE7792-EB58-493B-9098-791C8AF656FC}.Debug|x64.Build.0 = Debug|Any CPU
-		{AFFE7792-EB58-493B-9098-791C8AF656FC}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{AFFE7792-EB58-493B-9098-791C8AF656FC}.Debug|x86.Build.0 = Debug|Any CPU
-		{AFFE7792-EB58-493B-9098-791C8AF656FC}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{AFFE7792-EB58-493B-9098-791C8AF656FC}.Release|Any CPU.Build.0 = Release|Any CPU
-		{AFFE7792-EB58-493B-9098-791C8AF656FC}.Release|x64.ActiveCfg = Release|Any CPU
-		{AFFE7792-EB58-493B-9098-791C8AF656FC}.Release|x64.Build.0 = Release|Any CPU
-		{AFFE7792-EB58-493B-9098-791C8AF656FC}.Release|x86.ActiveCfg = Release|Any CPU
-		{AFFE7792-EB58-493B-9098-791C8AF656FC}.Release|x86.Build.0 = Release|Any CPU
 		{47D18D7C-1498-4318-A466-5A5CE6B80327}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{47D18D7C-1498-4318-A466-5A5CE6B80327}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{47D18D7C-1498-4318-A466-5A5CE6B80327}.Debug|x64.ActiveCfg = Debug|Any CPU
